Update database settings and add new URLs and views
This commit updates the database settings in the `config/settings.py` file to use SQLite3 as the default database. It also adds new URLs and views in the `config/urls.py` and `posts/views.py` files respectively. The new URLs include routes for Allauth, the home page, and serving media files. The new view `timeline` renders a timeline of posts in the `posts/timeline.html` template.

New files:
- Added `templates/base_generic.html` template file.
- Added `templates/posts/timeline.html` template file.